% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand2_main.R
\name{blocksize}
\alias{blocksize}
\title{Get a suitable blocksize for a given number of subjects}
\usage{
blocksize(x, N)
}
\arguments{
\item{x}{number of patients in stratum}

\item{N}{number of treatments}
}
\value{
integer vector of length one representing the blcksize
}
\description{
Get a suitable blocksize for a given number of subjects
}
\examples{
# Blocksize for 10 subjects and 2 treatments
blocksize(10, 2)

# Blocksize for 10 subjects and 3 treatments
blocksize(10, 3)
}

