% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand2_main.R
\name{createCombinations}
\alias{createCombinations}
\title{Create all possible combinations of values}
\usage{
createCombinations(strataVars)
}
\arguments{
\item{strataVars}{List of vectors containing the values to be combined. The
names of the list elements will appear as column names in the output data
frame.}
}
\value{
Data frame with each row representing a possible combination of
  values.
}
\description{
Create all possible combinations of values given in a list of value vectors
}
\examples{
createCombinations(list(
  sex = c("male", "female"),
  smokes = c("yes", "sometimes", "no"),
  cancer = c("yes", "no")
))
}

