% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand2_create.R
\name{createRandomisationDoc}
\alias{createRandomisationDoc}
\title{Create block randomisation list documents}
\usage{
createRandomisationDoc(patients, strataVars, treatments, header = list(title =
  "Randomisierungsliste zur Studie: title", author = "author", date =
  Sys.Date()), format = c("html", "word", "pdf"), newpage = FALSE,
  file.rmd = system.file("extdata", "randomList.Rmd", package = "blockrand2"),
  outdir = file.path(tempdir(), "blockrand2"))
}
\arguments{
\item{patients}{data frame containing the patient's data as returned by
\code{\link{createTestdata}}}

\item{strataVars}{List of strata variables, each of which is defined in
the form of a vector of possible (character) values.}

\item{treatments}{named vector of character defining the possible treatments.
The names are used as acronyms, the values as the full treatment names}

\item{header}{list with elements \code{title, author, date} defining the values to
be used in the title, author and date field, respectively, in the header of
the document}

\item{format}{vector of elements from \code{c("html", "word", "pdf")}
defining the types of documents to be created}

\item{newpage}{if \code{TRUE} and \code{"pdf"} is in \code{format} each
stratum will appear on its own page in the created PDF document}

\item{file.rmd}{full path to the RMarkdown file that is rendered to create
the randomisation list documents. By default the file \code{randomList.Rmd}
in the subfolder \code{extdata} of the package installation directory
(returned by \code{system.path(package = "blockrand2")} is used. You may
want to make a copy of this file, modify it and set \code{file.rmd} to the
path of that new file to override the appearence of the created output.}

\item{outdir}{full path to the output directory to which the created document
files are written. By default the files are written to a subfolder
\code{blockrand2} in the current R session's temporary directory}
}
\value{
list containing the full paths to the created files as values and the
  file extensions (\code{html, docx, pdf}) as element names
}
\description{
Create block randomisation list documents in three possible formats:
HTML, PDF (requires LaTeX) and Microsoft Word
}
\examples{
\dontrun{

# Define title of the study, the author and the date
header <- list(
  title = "Randomisation lists for the study: *Jokes against craziness*",
  author = "Hauke Sonnenberg",
  date = "2016-08-27"
)

# Define stratum variables
strataVars <- list(
  sex = c("male", "female"),
  crazyness = c("weak", "medium", "strong")
)

# Define the treatments
treatments <- c(
  joke = "Tell funny jokes",
  nojoke = "Keep serious!"
)

# Create some patient's testdata using the stratum levels
patients <- createTestdata(strataVars, 40)

# Create the randomisation list documents. We store the paths to the created
# files that are returned invisibly by createRandomisationDoc() in a variable
files <- createRandomisationDoc(
  patients = patients,
  strataVars = strataVars,
  treatments = treatments,
  header = header,
  newpage = TRUE
)

# Show the paths of the created files
files

# Open the html file in the default browser
browseURL(files$html)

# Open the pdf file in the default PDF viewer
system(paste(getOption("pdfviewer"), files$pdf))
}
}
\seealso{
\url{https://github.com/hsonne/blockrand2} (provides a short
  tutorial in the README file)
}

