% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand2_main.R
\name{createRandomSequences}
\alias{createRandomSequences}
\title{Create random sequences of blockwise equally distributed values}
\usage{
createRandomSequences(x, counts, names.strata = NULL)
}
\arguments{
\item{x}{vector of values to be chosen from}

\item{counts}{vector of integers defining the lengths of vectors of values to
be created}

\item{names.strata}{vector of character to be used as element names in the
output list}
}
\value{
list of vectors of values out of \code{x} with lengths according to
 the values given in \code{counts}
}
\description{
Create random sequences of blockwise equally distributed values. The
blocksizes are chosen according to the number of values to be created.
}
\examples{
# Define stratum levels
strata <- list(
 sex = c("male", "female"),
 medication = c("yes", "no"),
 lesion = c("low", "high")
)

# Create some testdata using the stratum levels and group and count by stratum
byStratum <- stratify(createTestdata(strata, 50))

# Create random sequences of treatments (A, B)
sequences <- createRandomSequences(
 x = c("A", "B"),
 counts = byStratum$n,
 names.strata = byStratum$stratum
)

# Append a column "treatments" showing the sequences of treatments as comma
# separated lists
byStratum$treatments <- sapply(sequences, paste, collapse = ",")

# Show the result
byStratum
}

