% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand2_main.R
\name{createTestdata}
\alias{createTestdata}
\title{Create testdata for clinical studies}
\usage{
createTestdata(strata, n = 30, format.patient = "P\%03d")
}
\arguments{
\item{strata}{List of strata}

\item{n}{Number of records to be created}

\item{format.patient}{format string to be used in \code{format} to create a
unique id for each patient}
}
\value{
Data frame with a column \code{patient} containing the patient's ID
and one column for each stratum, named according to the element names in
\code{strata}
}
\description{
Create testdata for clinical studies
}
\examples{
createTestdata(
  strata = list(
    sex = c("female", "male"),
    medication = c("yes", "no")
  ),
  n = 20
)
}

