% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand2_main.R
\name{equalDistribution}
\alias{equalDistribution}
\title{Create a random sequence of equally distributed values}
\usage{
equalDistribution(x, length.out = length(x), do.stop = TRUE)
}
\arguments{
\item{x}{Vector of values to be chosen from}

\item{length.out}{Length of the output vector to be created}

\item{do.stop}{If \code{TRUE} (default) an error is thrown if
\code{length.out} is not a multiple of the length of \code{x}. Otherwise
no error is thrown.}
}
\value{
Vector of \code{length.out} values all of which are elements of
  \code{x}. If \code{length.out} is a multiple of the lenght of \code{x} the
  values are equally distributed. Otherwise (and if \code{do.stop} is
  \code{FALSE} so that no error is thrown) the frequencies of the values
  differ by one at most.
}
\description{
Create a random sequence of equally distributed values
}
\examples{
y1 <- equalDistribution(x = LETTERS[1:2], 10)

# Check the distribution with table
table(y1)

# Do all values occur with the same frequency?
all(diff(table(y1)) == 0)

y2 <- equalDistribution(x = LETTERS[1:3], length.out = 11, do.stop = FALSE)

# Do the frequencies differ by one at most?
all(diff(table(y2)) <= 1)
}

