% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand2_main.R
\name{toStratumString}
\alias{toStratumString}
\title{Create a string describing the stratum from a one-row data frame}
\usage{
toStratumString(stratumInfo)
}
\arguments{
\item{stratumInfo}{data frame with one row and stratum information in all but
the very first (\code{stratum}) and the very last (\code{n}) column.}
}
\value{
character string of the form attribute1 = value1, attribute2 = value2,
 etc.
}
\description{
Create a string describing the stratum from a one-row data frame
}
\examples{
stratumInfo <- data.frame(
 stratum = "S01",
 sex = "male",
 pretreatment = "no",
 n = 5
)

toStratumString(stratumInfo)
}

