% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand2_main.R
\name{stratify}
\alias{stratify}
\title{Group and count data by combinations of strata}
\usage{
stratify(x, strata = toStrataList(x, exclude = names(x)[1]), column.n = "n",
  format.stratum = "S\%02d")
}
\arguments{
\item{x}{Data frame with stratum values in columns}

\item{strata}{List of strata. By default it is created from all but the first
columns of \code{x}.}

\item{column.n}{Column name to be used in the output for the column
containing the number of records falling into the combination of stratum
values.}

\item{format.stratum}{format string to be used in \code{format} to create a
unique id for each stratum}
}
\value{
Data frame with each row representing a combination of stratum values
}
\description{
Group data by combinations of stratum values and count the records falling
into each combination.
}
\examples{
strata <- list(sex = c("female", "male"), medication = c("yes", "no"))
x <- createTestdata(strata)
stratify(x, strata)
stratify(x, strata, column.n = "n.patients")
}

